refactor(catalog): move UserFilterFavorite to Catalog bundle (VIEW-27 fix #564)
Deptrac caught Identity_Internals → Catalog_Internals on the entity's
Attribute association. Resolution: favorites are catalog-side keyed
by Attribute, so the whole stack (entity, repo, controller, mapping)
moves into Catalog bundle. user_id stays as raw Uuid (mirror of the
SavedView / SmartFilterPreset pattern); Security principal is read
via the Shared UserIdentityAware contract so Catalog never imports
the Identity User entity.

Files moved:
- App\Identity\Domain\Entity\UserFilterFavorite -> App\Catalog\Domain\Entity\UserFilterFavorite
- App\Identity\Domain\Repository\UserFilterFavoriteRepositoryInterface -> App\Catalog
- App\Identity\Infrastructure\Doctrine\Repository\DoctrineUserFilterFavoriteRepository -> App\Catalog
- App\Identity\Infrastructure\Doctrine\Orm\Mapping\UserFilterFavorite.orm.xml -> App\Catalog
- App\Identity\Presentation\UserFilterFavoritesController -> App\Catalog\Presentation\Controller

PHPStan + Deptrac green; GET /api/users/me/filter-favorites returns
{"favorites":[]}.

Refs #558
