+++ apps/api/src/Catalog/Domain/Repository/UserFilterFavoriteRepositoryInterface.php
@@ -2,16 +2,19 @@
 
 declare(strict_types=1);
 
-namespace App\Identity\Domain\Repository;
+namespace App\Catalog\Domain\Repository;
 
-use App\Identity\Domain\Entity\User;
-use App\Identity\Domain\Entity\UserFilterFavorite;
+use App\Catalog\Domain\Entity\UserFilterFavorite;
+use Symfony\Component\Uid\Uuid;
 
 /**
  * VIEW-27 (#558) — repository contract for the per-user favorite
  * attribute shortcut list. Composite PK so the typical `find()` shape
  * does not apply; call sites either fetch the full list for a user
  * (picker) or atomically replace it (PUT).
+ *
+ * `user_id` is passed as a raw UUID rather than a User entity so the
+ * repository (Catalog bundle) does not pull in Identity_Internals.
  */
 interface UserFilterFavoriteRepositoryInterface
 {
@@ -21,16 +24,16 @@ interface UserFilterFavoriteRepositoryInterface
      *
      * @return list<UserFilterFavorite>
      */
-    public function findByUser(User $user): array;
+    public function findByUser(Uuid $userId): array;
 
     /**
-     * Atomic replace of the entire list for a user. Wipes existing
-     * rows and inserts the supplied (attribute, sort_order) tuples in
-     * the same transaction.
+     * Atomic replace of the entire list for a user. Wipes existing rows
+     * and inserts the supplied (attribute, sort_order) tuples in the
+     * same transaction.
      *
      * @param list<array{attribute_id: string, sort_order: int}> $entries
      */
-    public function replaceForUser(User $user, array $entries): void;
+    public function replaceForUser(Uuid $userId, array $entries): void;
 
     public function save(UserFilterFavorite $favorite): void;
 }
