+++ apps/api/src/Catalog/Presentation/Controller/UserFilterFavoritesController.php
